requirements.txt
gevent==1.0
python-gflags==2.0
requests==2.2.1
